fix: resolve MSVC build failures for Windows installer
- Upgrade spdlog CPM dependency 1.8.2 → 1.12.0 (fmt v7 in 1.8.2 has
  known incompatibilities with MSVC 2022 in C++20 mode)
- Cast GstMessageType enum to int before passing to spdlog fmt (fmt v9
  deprecated implicit enum formatting; MSVC treats this as an error)
- Add /verbosity:minimal /clp:ErrorsOnly to MSBuild invocation so
  the actual compiler errors are visible in CI logs on future failures

+++ src/audiorecorder.cpp
@@ -122,7 +122,7 @@ void AudioRecorder::processGstMessage() {
                     logger->debug("{} [gstreamer] Unhandled GStreamer msg received - Element: {} - Type: {} - Name: {}",
                                   m_loggingPrefix,
                                   message->src->name,
-                                  GST_MESSAGE_TYPE(message),
+                                  static_cast<int>(GST_MESSAGE_TYPE(message)),
                                   GST_MESSAGE_TYPE_NAME(message)
                     );
             }
